car_api/pyproject.toml
[tool.poetry]
name = "car-api"
version = "0.1.0"
description = ""
authors = ["Anhelina Sadvari <sadvari.anhelina@student.uzhnu.edu.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
faker = "^30.1.0"
faker-vehicle = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
